backend/requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==4.1.7
django-filter==23.1
djangorestframework==3.14.0
filelock==3.10.7
Markdown==3.4.3
platformdirs==3.2.0
pytz==2023.3
sqlparse==0.4.3
virtualenv==20.21.0
